EMit correct thing

+++ NERODevelopment/src/controllers/speedcontroller.h
@@ -25,7 +25,7 @@ class SpeedController : public ButtonController {
       int lastTime READ lastTime WRITE setLastTime NOTIFY lastTimeChanged)
   Q_PROPERTY(int currentSpeed READ currentSpeed WRITE setCurrentSpeed NOTIFY
                  currentSpeedChanged)
-  Q_PROPERTY(int regen READ regen WRITE setRegen NOTIFY regenChanged)
+  Q_PROPERTY(float regen READ regen WRITE setRegen NOTIFY regenChanged)
   Q_PROPERTY(
       int maxSpeed READ maxSpeed WRITE setMaxSpeed NOTIFY maxSpeedChanged)
   Q_PROPERTY(float current READ current WRITE setCurrent NOTIFY currentChanged)
@@ -52,7 +52,7 @@ public:
   float maxCurrent() const;
   float currentDischarge() const;
   float maxCurrentDischarge() const;
-  int regen() const;
+  float regen() const;
 
 signals:
   void tractionControlChanged(bool);
@@ -68,7 +68,7 @@ signals:
   void maxCurrentChanged(float);
   void currentDischargeChanged(float);
   void maxCurrentDischargeChanged(float);
-  void regenChanged(int);
+  void regenChanged(float);
 
 public slots:
   void setTractionControl(bool);
@@ -84,7 +84,7 @@ public slots:
   void setMaxCurrent(float);
   void setCurrentDischarge(float);
   void setMaxCurrentDischarge(float);
-  void setRegen(int);
+  void setRegen(float);
 
   void enterButtonPressed() override;
   void updateCurrentTime();
@@ -105,7 +105,7 @@ private:
   float m_maxCurrent = 0;
   float m_currentDischarge = 0;
   float m_maxCurrentDischarge = 0;
-  int m_regen = 0;
+  float m_regen = 0;
 
   bool m_timerRunning = false;
   QElapsedTimer m_timer;
